Facebook Login works / Console output is edited
The console was cluttered with OS messages, so I tampered with the console settings so we only see what we want. Facebook Login works now, but I've yet to test if we can extract user data properly (should be fine though). I also still need to add validation for FB logins.